+++ src/KT0803.cpp
@@ -60,12 +60,15 @@ bool KT0803::setChannel(uint16_t channel)
   //  need to split over 3 registers
   //  register 2 part skipped (always 0) for KT0803
   uint16_t ch = channel >> 1;
-  //  register 0
-  if (writeData(0x00, ch & 0xFF) == false) return false;
-  //  register 1
-  uint8_t data = readData(0x01) & 0xF8;  //  keep other bits
-  data |= ((ch >> 8) & 0x07);
-  return writeData(0x01, data);
+
+  uint8_t register0 = ch & 0xFF;  //  CHSEL[8:1]
+  if (writeData(0x00, register0) == false) return false;
+
+  ch >>= 8;
+  uint8_t register1 = readData(0x01);
+  register1 &= 0xF8;         //  CHSEL[11:9]
+  register1 |= (ch & 0x07);  //  CHSEL[11:9]
+  return writeData(0x01, register1);
 }
 
 
@@ -232,19 +235,20 @@ bool KT0803K::setChannel(uint16_t channel)
   if ((channel < 1400) || (channel > 2160)) return false;
   //  need to split over 3 registers
   uint16_t ch = channel;
-  //  register 2
-  uint8_t data = readData(0x02) & 0x7F;
-  data |= (channel & 0x01) << 8;
-  if (writeData(0x02, data) == false) return false;
+
+  uint8_t register2 = readData(0x02) & 0x7F;
+  register2 |= (channel & 0x01) << 7;   //  CHSEL[0]
+  if (writeData(0x02, register2) == false) return false;
   ch >>= 1;
-  //  register 0
-  if (writeData(0x00, ch & 0xFF) == false) return false;
-  //  register 1
+
+  uint8_t register0 = ch & 0xFF;  //  CHSEL[8:1]
+  if (writeData(0x00, register0) == false) return false;
+
   ch >>= 8;
-  data = readData(0x01);
-  data &= 0xF8;   //  keep other bits
-  data |= ch & 0x07;
-  return writeData(0x01, data);
+  uint8_t register1 = readData(0x01);
+  register1 &= 0xF8;         //  CHSEL[11:9]
+  register1 |= (ch & 0x07);  //  CHSEL[11:9]
+  return writeData(0x01, register1);
 }
 
 uint16_t KT0803K::getChannel()
@@ -253,7 +257,7 @@ uint16_t KT0803K::getChannel()
   channel <<= 8;
   channel |= readData(0x00);
   channel <<= 1;
-  channel |= (readData(0x02) >> 0x07);
+  channel |= (readData(0x02) >> 7);
   return channel;
 }